# openapi.yaml
openapi: 3.0.3
info:
  title: IM Backend API
  version: 0.1.0
paths:
  /auth/register:
    post:
      operationId: registerUser
      summary: Register a user and create the first device session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User and first device session created
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
            Cache-Control:
              $ref: '#/components/headers/PrivateNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Malformed JSON or invalid registration data
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterBadRequest'
        '409':
          description: Username is already registered
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterConflict'
        '429':
          description: Registration rate limit exceeded
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
            Retry-After:
              description: Whole seconds until another attempt is allowed
              required: true
              schema:
                type: string
                pattern: '^[1-9][0-9]*$'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthRateLimited'
        '500':
          description: Unexpected registration failure
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
        '503':
          description: Authentication rate limiter is temporarily unavailable
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthRateLimitUnavailable'
  /auth/login:
    post:
      operationId: login
      summary: Authenticate and idempotently create or recover a device session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated session created or recovered
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
            Cache-Control:
              $ref: '#/components/headers/PrivateNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Malformed JSON or invalid login request metadata
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginBadRequest'
        '401':
          description: Username or password is invalid
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidCredentials'
        '409':
          description: loginRequestId belongs to another login
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginConflict'
        '429':
          description: Login rate limit exceeded
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
            Retry-After:
              description: Whole seconds until another attempt is allowed
              required: true
              schema:
                type: string
                pattern: '^[1-9][0-9]*$'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthRateLimited'
        '500':
          description: Unexpected login failure
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
        '503':
          description: Login recovery or authentication rate limiter is temporarily unavailable
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginUnavailable'
  /auth/refresh:
    post:
      operationId: refreshSession
      summary: Rotate a refresh token or recover the same rotation result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Session tokens rotated or idempotently recovered
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
            Cache-Control:
              $ref: '#/components/headers/PrivateNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Malformed JSON or invalid idempotency key
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshBadRequest'
        '401':
          description: Refresh token is invalid, expired, revoked, or replayed with a different key
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenInvalid'
        '429':
          description: Refresh rate limit exceeded
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
            Retry-After:
              description: Whole seconds until another attempt is allowed
              required: true
              schema:
                type: string
                pattern: '^[1-9][0-9]*$'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthRateLimited'
        '503':
          description: Refresh recovery or authentication rate limiter is temporarily unavailable
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshUnavailable'
  /auth/logout:
    post:
      operationId: logout
      summary: Revoke the current device session
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Current session revoked
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
        '401':
          description: A valid access token is required
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationRequired'
        '500':
          description: Unexpected logout failure
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
        '503':
          description: Authentication dependency is temporarily unavailable
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationUnavailable'
  /auth/logout-all:
    post:
      operationId: logoutAll
      summary: Revoke all sessions for the authenticated user
      security:
        - bearerAuth: []
      responses:
        '204':
          description: All sessions revoked
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
        '401':
          description: A valid access token is required
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationRequired'
        '500':
          description: Unexpected logout failure
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
        '503':
          description: Authentication dependency is temporarily unavailable
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationUnavailable'
  /auth/sessions:
    get:
      operationId: listSessions
      summary: List all device sessions for the authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Device sessions, including revoked sessions
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
        '401':
          description: A valid access token is required
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationRequired'
        '500':
          description: Unexpected session listing failure
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
        '503':
          description: Authentication dependency is temporarily unavailable
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationUnavailable'
  /auth/sessions/{sessionID}:
    delete:
      operationId: revokeSession
      summary: Revoke one active device session owned by the authenticated user
      security:
        - bearerAuth: []
      parameters:
        - name: sessionID
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '204':
          description: Session revoked
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
        '400':
          description: sessionID is not a positive integer
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidSessionID'
        '401':
          description: A valid access token is required
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationRequired'
        '404':
          description: Active owned session was not found
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionNotFound'
        '500':
          description: Unexpected session revocation failure
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
        '503':
          description: Authentication dependency is temporarily unavailable
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationUnavailable'
  /auth/password:
    post:
      operationId: changePassword
      summary: Change the password and revoke all device sessions
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Password changed and all sessions revoked
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
        '400':
          description: Malformed JSON or invalid new password
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangePasswordBadRequest'
        '401':
          description: Access token or current password is invalid
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangePasswordUnauthorized'
        '409':
          description: Password changed concurrently
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordChangeConflict'
        '500':
          description: Unexpected password change failure
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
        '503':
          description: Authentication dependency is temporarily unavailable
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationUnavailable'
...
  /conversations:
    get:
      operationId: listConversations
      summary: List a stable snapshot of the authenticated user's conversations
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/After'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/SnapshotCursor'
      responses:
        '200':
          description: Conversation snapshot page
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationListResponse'
        '400':
          description: Invalid pagination parameters
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidPagination'
        '401':
          description: A valid access token is required
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationRequired'
        '409':
          description: A cursor is ahead of the current conversation list
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CursorAhead'
        '500':
          description: Unexpected conversation listing failure
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageInternalError'
        '503':
          description: Authentication dependency is temporarily unavailable
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationUnavailable'
...
  /messages/sync:
    get:
      deprecated: true
      operationId: deprecatedUserMessageSync
      summary: Deprecated user-level message Sync endpoint
      security:
        - bearerAuth: []
      responses:
        '401':
          description: A valid access token is required
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationRequired'
        '410':
          description: Use conversation-scoped recovery endpoints
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndpointGone'
        '503':
          description: Authentication dependency is temporarily unavailable
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-token
  headers:
    RequestID:
      description: Server-generated identifier used to correlate the response with logs
      required: true
      schema:
        type: string
        minLength: 1
    PrivateNoStore:
      description: Prevent storage of token-bearing authentication responses
      required: true
      schema:
        type: string
        enum:
          - no-store
  parameters:
...
    After:
      name: after
      in: query
      required: false
      schema:
        type: integer
        format: int64
        minimum: 0
        default: 0
...
  schemas:
    RegisterRequest:
      type: object
      additionalProperties: false
      required:
        - username
        - displayName
        - password
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 32
          description: Normalized to lowercase and trimmed before server validation
        displayName:
          type: string
          minLength: 1
          maxLength: 100
        password:
          type: string
          format: password
          writeOnly: true
          minLength: 8
          maxLength: 128
        deviceId:
          type: string
          maxLength: 128
...
    RegisterBadRequest:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
          enum:
            - INVALID_JSON
            - INVALID_USERNAME
            - INVALID_DISPLAY_NAME
            - INVALID_PASSWORD
            - INVALID_DEVICE_ID
        message:
          type: string
          minLength: 1
        requestId:
          type: string
          minLength: 1
...